+++ eva/src/js/componentsSettings.js
@@ -337,6 +337,7 @@ export default {
     'visible',
     'level',
     'pinned',
+    'excludedFromPDF',
     'otherSettings',
   ],
   options: {
@@ -616,6 +617,12 @@ export default {
       option: 'pinned',
       description: 'Закрепить на всех вкладках',
       elem: 'switch',
+    },
+    {
+      option: 'excludedFromPDF',
+      description: 'Исключить из PDF',
+      elem: 'switch',
+      default: false,
     },
     {
       option: 'level',